review comments addressed except for class headers
