+++ gbpservice/nfp/service_vendor_agents/vyos/agent/DEBIAN/changelog
@@ -13,5 +13,3 @@
 vyos (2.0) INITIAL RELEASE; urgency=low
 
   * Initial release. (Closes: #XXXXXX)
-
- -- root <root@controller-desktop>  Wed, 13 May 2015 10:13:46 +0530
