+++ gbpservice/nfp/service_vendor_agents/vyos/agent/bin/vyos
@@ -16,5 +16,5 @@ import sys
 
 sys.path.append("/usr/share/vyos")
 
-from config_server.server2 import main
+from config_server.server import main
 main()
